requirements.txt
PyPDF2>=3.0.0
langchain>=0.0.267
sentence-transformers>=2.2.2
chromadb>=0.4.13
flask>=2.3.3
nltk>=3.8.1
scikit-learn>=1.3.0
python-dotenv>=1.0.0
openai>=0.28
streamlit>=1.20.0
